clean login commands
